refactor: lazy load dashboard module
